Fix bug when finding previous/next revision
The problem was that sometimes, the information in
`git-timemachine-revision` is not exactly the same as the information
found in `(git-timemachine--revisions). For example, the
`git-timemachine--revision`

  ("853c..." 110 "9 seconds ago" "Thu Mar 26 05:53:41 2015 +0000")

has not the same relative date as the first element of
`(git-timemachine--revisions)`:

  (("853c..." 110 "2 minutes ago" "Thu Mar 26 05:53:41 2015 +0000") ...)

The patch I propose makes sure that only the hash is compared, not the
meta-data.